pyproject.toml
[project]
name = "starknet-stakemate"
version = "2.0.0"
description = "Starknet staking monitoring bot + REST API + Mini App (Staking V2 ready)"
readme = "README.md"
requires-python = ">=3.14,<4"
license = { text = "MIT" }
authors = [{ name = "saniksin" }]
keywords = ["starknet", "staking", "telegram", "bot", "mini-app"]

dependencies = [
    # Telegram
    "aiogram>=3.27.0,<4.0",
    "aiogram-dialog>=2.6.0,<3.0",
    # Starknet
    "starknet-py>=0.30.0,<0.31",
    # HTTP/API
    "fastapi>=0.136.0,<1.0",
    "uvicorn[standard]>=0.46.0,<1.0",
    "httpx>=0.28.0,<1.0",
    # Data layer
    "sqlalchemy>=2.0.49,<3.0",
    "aiosqlite>=0.22.0,<1.0",
    # Core
    "pydantic>=2.11.0,<2.13",  # capped by aiogram 3.27 peer requirement
    "python-dotenv>=1.2.0,<2.0",
    "loguru>=0.7.3,<1.0",
    "aiocache>=0.12.3,<1.0",
    "tenacity>=9.1.0,<10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9.0",
    "pytest-asyncio>=0.24,<1.0",
    "pytest-cov>=5.0,<7.0",
    "ruff>=0.7,<1.0",
    "mypy>=1.13,<2.0",
]

[project.scripts]
stakemate-bot = "main:run"
stakemate-api = "api.app:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "bot",
    "data",
    "db_api",
    "parse",
    "services",
    "api",
    "tasks",
    "utils",
]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM"]
ignore = ["E501"]  # long lines tolerated in Telegram HTML message constants

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
